Fixed some small things and put parse_dict function from test file into segmentation_data_prep file
